requirements.txt

selenium
fpdf2
PIL
advertools
urllib
bs4
pandas
requests
pytz
SQLAlchemy
Scrapy
pyodbc
setuptools
pymongo
tld
pdfkit
requests
beautifulsoup4
PyPDF2
Twisted
